+++ packages/create-analog/template-angular-v19/src/main.ts
@@ -1,0 +1,7 @@
+import 'zone.js';
+import { bootstrapApplication } from '@angular/platform-browser';
+
+__APP_COMPONENT_IMPORT__
+import { appConfig } from './app/app.config';
+
+bootstrapApplication(__APP_COMPONENT__, appConfig);