+++ packages/nx-plugin/src/generators/init/lib/update-build-target.ts
@@ -34,11 +34,28 @@ export function updateBuildTarget(
   };
 
   if (tree.exists(angularJsonPath)) {
+    const projects = getProjects(tree);
+
+    const projectConfig = projects.get(schema.project);
+
     updateJson(tree, angularJsonPath, (json) => {
-      json.projects[schema.project].root = '.';
+      json.projects[schema.project].root = projectConfig.root;
+      json.projects[schema.project].sourceRoot = projectConfig.sourceRoot;
       json.projects[schema.project].architect.build = {
         builder: '@analogjs/platform:vite',
         ...commonConfig,
+        options: {
+          configFile: `${joinPathFragments(
+            projectConfig.root,
+            'vite.config.ts',
+          )}`,
+          main: `${joinPathFragments(projectConfig.root, 'src/main.ts')}`,
+          outputPath: `dist/${joinPathFragments(projectConfig.root, 'client')}`,
+          tsConfig: `${joinPathFragments(
+            projectConfig.root,
+            'tsconfig.app.json',
+          )}`,
+        },
       };
 
       return json;