+++ packages/vite-plugin-nitro/src/lib/plugins/dev-server-plugin.ts
@@ -23,7 +23,6 @@ type ServerOptions = Options & { routeRules?: Record<string, any> | undefined };
 export function devServerPlugin(options: ServerOptions): Plugin {
   const workspaceRoot = options?.workspaceRoot || process.cwd();
   const sourceRoot = options?.sourceRoot ?? 'src';
-  const entryServer = options.entryServer || `${sourceRoot}/main.server.ts`;
   const index = options.index || 'index.html';
   let config: UserConfig;
   let root: string;
@@ -38,7 +37,8 @@ export function devServerPlugin(options: ServerOptions): Plugin {
       return {
         resolve: {
           alias: {
-            '~analog/entry-server': entryServer,
+            '~analog/entry-server':
+              options.entryServer || `${root}/${sourceRoot}/main.server.ts`,
           },
         },
       };